+++ lib/enums/asuka_type_enum.dart
@@ -1,0 +1,4 @@
+enum AsukaType {
+  snackbar,
+  materialBanner,
+}